requirements.txt

aiogram>=3.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog
python-dotenv
orjson
celery
redis
python-json-logger
prometheus-client
pymupdf>=1.23.0
pytesseract
pdf2image
pdfplumber
python-docx>=0.8.11
transliterate
python-Levenshtein
yadisk>=1.3.4
aiohttp
aiofiles
pytest
pytest-asyncio
ocrmypdf>=15.4.0
pillow>=10.0.0
requests
aioresponses

PyPDF2>=3.0.0
